LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/mixin-from-import-expected.txt

This text should be green.

FAIL CSS Mixins: Mixin from an imported stylesheet assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
